CRM-2109: RFM analytic for Magento channels - fixed code analys software warnings

+++ src/OroCRM/Bundle/AnalyticsBundle/Model/RFMMetricStateManager.php
@@ -108,7 +108,7 @@ class RFMMetricStateManager
      */
     public function scheduleRecalculation(Channel $channel = null, $flush = true)
     {
-        if ($job = $this->getJob()) {
+        if ($this->getJob()) {
             return;
         }
 
@@ -129,8 +129,10 @@ class RFMMetricStateManager
         if (!$channel) {
             $channelJobs = $this->getJob('--channel');
 
-            foreach ($channelJobs as $channelJob) {
-                $em->remove($channelJob);
+            if ($channelJobs) {
+                foreach ($channelJobs as $channelJob) {
+                    $em->remove($channelJob);
+                }
             }
         }
